API view that is function and class based and mixins get introduced

+++ checkapi/views.py
@@ -9,17 +9,95 @@ from rest_framework.parsers import JSONParser
 from rest_framework.decorators import api_view 
 from rest_framework.response import Response
 from rest_framework import status 
+from rest_framework.views import APIView
 
 from django.views.decorators.csrf import csrf_exempt
 from django.utils.decorators import method_decorator
 from django.views import View
 
+from rest_framework.generics import GenericAPIView
+from rest_framework.mixins import ListModelMixin, CreateModelMixin, UpdateModelMixin, RetrieveModelMixin, DestroyModelMixin
+
+
 # Create your views here.
 
-@api_view(['GET', 'POST', 'PUT', 'DELETE'])
-def student(request):
+# List and create
+class LCStudentList(GenericAPIView, ListModelMixin, CreateModelMixin, UpdateModelMixin, RetrieveModelMixin, DestroyModelMixin):
+    queryset = Student.objects.all()
+    serializer_class = StudentSerializer 
+
+    def get(self, request, *args, **kwargs):
+        return self.list(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        return self.create(request, *args, **kwargs)
+
+class URDStudent(GenericAPIView, RetrieveModelMixin, UpdateModelMixin, DestroyModelMixin):
+    queryset = Student.objects.all()
+    serializer_class = StudentSerializer
+
+    def get(self, request, *args, **kwargs):
+        return self.retrieve(request, *args, **kwargs)
+
+    def put(self, request, *args, **kwargs):
+        return self.update(request, *args, **kwargs)
+    
+    def patch(self, request, *args, **kwargs):
+        return self.partial_update(request, *args, **kwargs)
+    
+    def delete(self, request, *args, **kwargs):
+        return self.destroy(request, *args, **kwargs)
+
+
+class StudentAPI(APIView):
+    def get(self, request, id=None, format=None):
+        if id is not None:
+            stu = Student.objects.get(id=id)
+            serializer = StudentSerializer(stu)
+            return Response(serializer.data)
+        stu = Student.objects.all()
+        serializer = StudentSerializer(stu, many=True)
+        return Response(serializer.data)
+    
+    def post(self, request, format=None):
+        serializer = StudentSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    def put(self, request, id=None, format=None):
+        stu = Student.objects.get(id=id)
+        serializer = StudentSerializer(stu, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({
+                'msg':'data updated'
+            })
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    def patch(self, request, id=None, format=None):
+        stu = Student.objects.get(id=id)
+        serializer = StudentSerializer(stu, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({
+                'msg':'data patch updated'
+            })
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+    def delete(self, request, id=None, format=None):
+        stu = Student.objects.get(id=id)
+        stu.delete()
+        return Response({
+            'msg':'data deleted'
+        })
+        
+
+@api_view(['GET', 'POST', 'PUT', 'PATCH', 'DELETE'])
+def student(request, id = None):
     if request.method == 'GET':
-        id = request.GET.get('id', None)
+        # id = request.GET.get('id', None)
         if id is not None:
             stu = Student.objects.get(id=id)
             serializer = StudentSerializer(stu)
@@ -36,9 +114,9 @@ def student(request):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     if request.method == 'PUT':
-        id = request.data.get('id', None)
+        # id = request.data.get('id', None)
         stu = Student.objects.get(id=id)
-        serializer = StudentSerializer(stu, data=request.data, partial=True)
+        serializer = StudentSerializer(stu, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response({
@@ -46,8 +124,19 @@ def student(request):
             })
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    if request.method == 'PATCH':
+        # id = request.data.get('id', None)
+        stu = Student.objects.get(id=id)
+        serializer = StudentSerializer(stu, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response({
+                'msg':'data patch updated'
+            })
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
     if request.method == 'DELETE':
-        id = request.data.get('id', None)
+        # id = request.data.get('id', None)
         stu = Student.objects.get(id=id)
         stu.delete()
         return Response({
@@ -70,72 +159,72 @@ def student_create(request):
         return HttpResponse(json_data, content_type='application/json')
         
      
-@method_decorator(csrf_exempt, name='dispatch')
-class StudentAPI(View):
-    def get(self, request, *args, **kwargs):
-        json_data = request.body
-        stream = io.BytesIO(json_data)
-        py_data = JSONParser().parse(stream)
-        id = py_data.get('id', None)
+# @method_decorator(csrf_exempt, name='dispatch')
+# class StudentAPI(View):
+#     def get(self, request, *args, **kwargs):
+#         json_data = request.body
+#         stream = io.BytesIO(json_data)
+#         py_data = JSONParser().parse(stream)
+#         id = py_data.get('id', None)
         
         
-        if id is not None:
-            stu = Student.objects.get(id=id)
-            serializer = StudentSerializer(stu)
-            json_data = JSONRenderer().render(serializer.data)
-            return HttpResponse(json_data, content_type='application/json')
+#         if id is not None:
+#             stu = Student.objects.get(id=id)
+#             serializer = StudentSerializer(stu)
+#             json_data = JSONRenderer().render(serializer.data)
+#             return HttpResponse(json_data, content_type='application/json')
         
-        stu = Student.objects.all()
-        serializer = StudentSerializer(stu, many=True)
-        json_data = JSONRenderer().render(serializer.data)
-        return HttpResponse(json_data, content_type='application/json')
+#         stu = Student.objects.all()
+#         serializer = StudentSerializer(stu, many=True)
+#         json_data = JSONRenderer().render(serializer.data)
+#         return HttpResponse(json_data, content_type='application/json')
 
-    def post(self, request, *args, **kwargs):
-        json_data = request.body
-        stream = io.BytesIO(json_data)
-        py_data = JSONParser().parse(stream)
-        serializer = StudentSerializer(data=py_data)
+#     def post(self, request, *args, **kwargs):
+#         json_data = request.body
+#         stream = io.BytesIO(json_data)
+#         py_data = JSONParser().parse(stream)
+#         serializer = StudentSerializer(data=py_data)
         
-        if serializer.is_valid():
-            serializer.save()
-            res = {'msg':'data created'}
-            json_data = JSONRenderer().render(res)
-            return HttpResponse(json_data, content_type='application/json')
-        json_data = JSONRenderer().render(serializer.errors)
-        return HttpResponse(json_data, content_type='application/json')
+#         if serializer.is_valid():
+#             serializer.save()
+#             res = {'msg':'data created'}
+#             json_data = JSONRenderer().render(res)
+#             return HttpResponse(json_data, content_type='application/json')
+#         json_data = JSONRenderer().render(serializer.errors)
+#         return HttpResponse(json_data, content_type='application/json')
     
     
-    def put(self, request, *args, **kwargs):
-        json_data = request.body
-        stream = io.BytesIO(json_data)
-        py_data = JSONParser().parse(stream)
-        id = py_data.get('id')
+#     def put(self, request, *args, **kwargs):
+#         json_data = request.body
+#         stream = io.BytesIO(json_data)
+#         py_data = JSONParser().parse(stream)
+#         id = py_data.get('id')
         
-        stu = Student.objects.get(id=id)
-        serializer = StudentSerializer(stu, data=py_data, partial=True)
+#         stu = Student.objects.get(id=id)
+#         serializer = StudentSerializer(stu, data=py_data, partial=True)
         
-        if serializer.is_valid():
-            serializer.save()
-            res = {'msg':'data updated'}
-            json_data = JSONRenderer().render(res)
-            return HttpResponse(json_data, content_type='application/json')
+#         if serializer.is_valid():
+#             serializer.save()
+#             res = {'msg':'data updated'}
+#             json_data = JSONRenderer().render(res)
+#             return HttpResponse(json_data, content_type='application/json')
         
-        json_data = JSONRenderer().render(serializer.errors)
-        return HttpResponse(json_data, content_type='application/json')
+#         json_data = JSONRenderer().render(serializer.errors)
+#         return HttpResponse(json_data, content_type='application/json')
     
     
-    def delete(self, request, *args, **kwargs):
-        json_data = request.body
-        stream = io.BytesIO(json_data)
-        py_data = JSONParser().parse(stream)
-        id = py_data.get('id')
+#     def delete(self, request, *args, **kwargs):
+#         json_data = request.body
+#         stream = io.BytesIO(json_data)
+#         py_data = JSONParser().parse(stream)
+#         id = py_data.get('id')
 
-        stu = Student.objects.get(id=id)
-        stu.delete()
-        res = {'msg':'data deleted'}
-        # json_data = JSONRenderer().render(res)
-        # return HttpResponse(json_data, content_type='application/json')
-        return JsonResponse(res, safe=False)
+#         stu = Student.objects.get(id=id)
+#         stu.delete()
+#         res = {'msg':'data deleted'}
+#         # json_data = JSONRenderer().render(res)
+#         # return HttpResponse(json_data, content_type='application/json')
+#         return JsonResponse(res, safe=False)
